+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- Support for container types: `QSet<T>`, `QHash<K, V>`, `Qt5List<T>`, `QMap<K, V>`, `QVector<T>`
+- Support for container types: `QSet<T>`, `QHash<K, V>`, `QList<T>`, `QMap<K, V>`, `QVector<T>`
 - Support for further types: `QByteArray`, `QModelIndex`, `QPersistentModelIndex`
 - Support for nesting objects in properties, invokables, and signals with `*mut T`